Access titles & images w/ UISegmentedControl accessors
From the beginning, We added the following messages to
UISegmentedControl:

    -[UISegmentedControl insertSegmentWithTitle:image:atIndex:animated:]
    -[UISegmentedControl setTitle:image:forSegmentAtIndex:]

It allows us to have a UISegmentedControl present both a title string
*and* an image. Now, if we want to retrieve those title and image
components, we'll use UISegmentedControl's built in accessors:

    -[UISegmentedControl titleForSegmentAtIndex:]
    -[UISegmentedControl imageForSegmentAtIndex:]

Since we're still developing this title-with-image support as a
UISegmentedControl category, the above two methods are swizzled to
augment their behavior. If the segment of interest has a solitary title
or image, we'll use the default UISegmentedControl behavior. If not,
we'll retrieve either the title or image that was set via our two
additional methods.

At first glance this seems dangerous, and it may well be. We're going to
try and maintain decent test coverage and always be on the lookout for
wonkiness. When starting out on this project, the category strategy was
preferable over the subclassing strategy because it allowed us to add
the title-with-image behavior to *any* UISegmentedControl without
requiring consumers to use a different superclass. I think this is still
a worthy goal.